Add support for arbitrary notifications
