+++ src/domain/session/toast/IToastCollection.ts
@@ -1,0 +1,6 @@
+import {ObservableArray} from "../../../observable";
+import type {BaseToastNotificationViewModel} from "./BaseToastNotificationViewModel";
+
+export interface IToastCollection {
+    toastViewModels: ObservableArray<BaseToastNotificationViewModel>;
+}
